Add buttons to VnrMessageArea
Add GtkHButtonBox to the layout. To show the area with buttons
you pass the stock id of the button and the callback to be called
if the button is clicked. A "cancel" button is included by default.
Introduce is_critical - if it is
TRUE, a GTK_STOCK_DIALOG_WARNING is displayed, otherwise
GTK_STOCK_DIALOG_INFO is displayed. get_visible is replaced with
is_critical, where we need to check not the visibility, but if
there is an error showed.